+++ pkgs/timezone/tool/refresh.sh
@@ -9,10 +9,10 @@ temp=$(mktemp -d -t tzdata-XXXXXXXXXX)
 
 pushd $temp > /dev/null
 
-# Fetch latest database
+echo "Fetching latest database..."
 curl https://data.iana.org/time-zones/tzdata-latest.tar.gz | tar -zx
 
-# Compile into zoneinfo files
+echo "Compiling into zoneinfo files..."
 mkdir zoneinfo
 zic -d zoneinfo africa antarctica asia australasia etcetera europe \
                 northamerica southamerica backward
@@ -27,6 +27,9 @@ pub run tool/encode_tzf --zoneinfo $temp/zoneinfo
 rm -r $temp
 
 # Create the source embeddings
-for x in latest latest_all latest_10y; do
-  pub run tool/encode_dart lib/data/$x.{tzf,dart}
+for scope in latest latest_all latest_10y; do
+  echo "Creating embedding: $scope..."
+  pub run tool/encode_dart lib/data/$scope.{tzf,dart}
 done
+
+dart format lib/data